requirements.txt
django>=2.2
django-ckeditor>=5.7,<6.0
html2text>=2020
